src/client/app/+user-management/+forgot-password/forgot-password.component.spec.ts
import { Observable } from '../../imports/test.imports';
import { ForgotPasswordComponent } from './forgot-password.component';

export function main() {
  describe('Forgot Password Component', () => {
    let mockUiConfig: any, mockUser: any;
    let componentUnderTest: ForgotPasswordComponent;

    beforeEach(() => {
      mockUiConfig = { get: () => { return Observable.of({ config: { someConfig: 'test' } }); } };
      mockUser = { forgotPassword: jasmine.createSpy('forgotPassword').and.returnValue(Observable.of({})) };
      componentUnderTest = new ForgotPasswordComponent(mockUser, mockUiConfig);
    });

    describe('ngOnInit()', () => {
      it('Grabs the component config and assigns to an instance variable', () => {
        componentUnderTest.ngOnInit();
        expect(componentUnderTest.config).toEqual({ someConfig: 'test' });
      });
    });

    describe('onSubmit()', () => {
      it('Submits a request for a reset password email', () => {
        componentUnderTest.onSubmit({ 'emailAddress': 'test@test.com' });
        expect(componentUnderTest.user.forgotPassword).toHaveBeenCalledWith({ 'emailAddress': 'test@test.com' });
        expect(componentUnderTest.successfullySubmitted).toEqual(true);
      });
    });

    describe('ngOnDestroy()', () => {
      it('unsubscribes from the UI config to prevent memory leakage', () => {
        let mockSubscription = { unsubscribe: jasmine.createSpy('unsubscribe') };
        let mockObservable = { subscribe: () => mockSubscription };
        mockUiConfig = { get: () => mockObservable };
        componentUnderTest = new ForgotPasswordComponent(mockUser, mockUiConfig);
        componentUnderTest.ngOnInit();
        componentUnderTest.ngOnDestroy();
        expect(mockSubscription.unsubscribe).toHaveBeenCalled();
      });
    });

  });
}
